Added mutantAnalysis.R
This function performs a mutant analysis for each single node of a network by exploring a large sample
of initial states and saves the results into a separate file for each mutant. The user can chose between
"overexp" and "null" types of perturbations to simulate gain- or loss-of-function mutants, respectively.